Work with v2.x.x of the Google Terraform Provider
The Google Terraform provider we're using is being deprecated. We are
two major versions behind (3.x.x is current).  Jumping to v2.x.x
includes some breaking schema changes for GKE resources that we use
in this module. Making minimum viable updates to get module compatible
with the v2.x.x provider.

The intent is to make these changes in a way that does not require
rebuilding existing clusters.