Remove variables as descriptors
We employ map variables for the lions share of configurations in this
module.  These map variables help collect like configurations with the
intent of helping the operator keep track of what's being configured
where.

Sadly, map variables don't offer much in the description department.
As a suppliment to the map variables we included individual variable
definitions for each key pair in each map variable.  The intent for
these individual variables was not to be used, but simply to provide a
description for each configurations purpose.

After some reflection this feels pretty confusing.  For now I'm going to
try removing the descriptor variables in favor of the maps and seeing
how that goes.