Remove tags
In the current state when you try and pass the lookup default to tags it
results in a Terraform crash.  We've never run into this before because
we've always had a required value for tags, the NAT VM's regional
network tag.

We're now moving away from NAT VM's in favor of Googles hosted cloud-nat
service.  Rather than maintain the argument with tweaks to support
having it not set, we're just going to remove it for now.  We shouldn't
need tags any time soon, and can always add support back.

By making this change, your node pool will be recreated when using the
version of GKE module that includes this change.